+++ src/core/prelude/object/convert/index.ts
@@ -7,7 +7,7 @@
  */
 
 import extend from 'core/prelude/extend';
-import { canParse } from 'core/prelude/object/const';
+import { canParse, isInvalidKey } from 'core/prelude/object/const';
 
 /** @see [[ObjectConstructor.trySerialize]] */
 extend(Object, 'trySerialize', (value, replacer?: JSONCb) => {
@@ -50,7 +50,7 @@ extend(Object, 'trySerialize', (value, replacer?: JSONCb) => {
 extend(Object, 'parse', (value, reviver?: JSONCb) => {
 	if (Object.isFunction(value)) {
 		reviver = value;
-		return (value) => Object.parse(value, reviver);
+		return (value) => Object.parse(value, wrapReviver(reviver));
 	}
 
 	if (Object.isString(value)) {
@@ -61,7 +61,7 @@ extend(Object, 'parse', (value, reviver?: JSONCb) => {
 		if (canParse.test(value)) {
 			try {
 				const
-					parsedVal = JSON.parse(value, reviver);
+					parsedVal = JSON.parse(value, wrapReviver(reviver));
 
 				if (Object.isNumber(parsedVal)) {
 					return parsedVal.isSafe() ? parsedVal : value;
@@ -77,4 +77,19 @@ extend(Object, 'parse', (value, reviver?: JSONCb) => {
 	}
 
 	return value;
+
+	function wrapReviver(fn: Nullable<JSONCb>) {
+		return (key, value) => {
+			if (isInvalidKey.test(key)) {
+				return;
+			}
+
+			if (Object.isFunction(fn)) {
+				value = fn(key, value);
+			}
+
+			return value;
+		};
+	}
+
 });